+++ clang/test/PCH/chain-openmp-threadprivate.cpp
@@ -9,8 +9,6 @@
 // RUN: %clang_cc1 -fopenmp -emit-llvm -chain-include %s -chain-include %s %s -o - | FileCheck %s -check-prefix=CHECK-TLS-1
 // RUN: %clang_cc1 -fopenmp -emit-llvm -chain-include %s -chain-include %s %s -o - | FileCheck %s -check-prefix=CHECK-TLS-2
 
-// REQUIRES: tls
-
 #if !defined(PASS1)
 #define PASS1